// lib/cartographer/model/RepositoryModel.ts
export interface RepositoryModel {
  files: string[];

  pages: string[];
  components: string[];
  libraries: string[];
  engines: string[];
  repositories: string[];
  events: string[];
  migrations: string[];
  tests: string[];
  docs: string[];

  apiRoutes: string[];
  layouts: string[];
  middleware: string[];
  hooks: string[];
  contexts: string[];
  workers: string[];
}